quiz/pyproject.toml
[project]
name = "quiz"
version = "0.1.0"
description = ""
authors = [
    {name = "Lupuna",email = "terragod427@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
